refactor: Os DTOs ProductCreateDTO e ProductUpdateDTO foram adaptados ao tipo de dado BigDecimal. Além disso, o OrderProductResponseDTO teve que mudar a lógica para inserir o campo totalPrice no campo, visto que o tipo de dado BigDecimal funciona de forma diferente com operações matemáticas

+++ back-end/src/main/java/org/uorderflow/dto/product/ProductUpdateDTO.java
@@ -3,11 +3,13 @@ package org.uorderflow.dto.product;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public record ProductUpdateDTO(
         @Size(max = 100) String name,
         @Size(max = 255) String description,
         @Size(max = 2048) String image,
-        @Positive Double price,
+        @Positive BigDecimal price,
         Long productCategoryId,
         Boolean isAvailable
 ) {
